Move extensions types to apps
